Style#initialize - dup definition before deleting options
This was causing errors if a processor expects the geometry option to be
defined. It worked the first time through but on subsequent attempts,
options[:geometry] is nil because of definition.delete(:geometry). Whether
this error happens seems to depend on the particulars of the server
environment.

See http://github.com/ng/paperclip-watermarking-app/issues#issue/2 for
some discussion.